third_party/blink/web_tests/platform/generic/http/tests/security/mixedContent/insecure-script-in-iframe-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-script.html' was loaded over HTTPS, but requested an insecure script 'http://example.test:8080/security/mixedContent/resources/script.js'. This content should also be served over HTTPS.
This test loads a secure iframe that loads an insecure script. We should trigger a mixed content callback even though the main frame is HTTP because the HTTPS frame's origin is contaminated with an insecure script.
